Added support for other operating systems. Added classes for certain projects, as well as tasks associated with these projects.

+++ src/main/java/ru/hzerr/util/Cleaner.java
@@ -3,7 +3,6 @@ package ru.hzerr.util;
 import org.apache.commons.io.FileUtils;
 import ru.hzerr.GradleOptions;
 import ru.hzerr.HLogger;
-import ru.hzerr.Helper;
 
 import java.io.File;
 import java.io.IOException;
@@ -11,17 +10,24 @@ import java.util.Objects;
 
 public class Cleaner {
 
-    public static void main(String[] args) throws IOException { cleanup(Helper.parse(args)); }
+    public static void main(String[] args) throws IOException {
+        cleanup(GradleOptions.getGradleOptions(args));
+    }
 
     public static void cleanup(GradleOptions options) throws IOException {
-        for (File file : Objects.requireNonNull(options.getFolderFile().listFiles())) {
-            if (!file.getName().equals(options.getProjectName())) {
-                if (file.isDirectory()) {
-                    FileUtils.deleteDirectory(file);
-                } else FileUtils.forceDelete(file);
+        File[] files = options.getFolderFile().listFiles();
+        boolean completed = false;
+        if (files != null) {
+            for (File file : files) {
+                if (!file.getName().equals(options.getProjectName())) {
+                    completed = true;
+                    if (file.isDirectory()) {
+                        FileUtils.deleteDirectory(file);
+                    } else FileUtils.forceDelete(file);
+                }
             }
-        }
-
-        HLogger.success("The cleaning was successful");
+            if (completed) HLogger.success("The cleaning was successful");
+            else HLogger.info("There's nothing to clean up");
+        } else HLogger.info("There's nothing to clean up");
     }
 }
